Seperated main from UCSBCurriculumSearch file, deleted SearchGui_Curriculum because it was copy of UCSBgui, and deleted old commented out code
